# api/openapi.yaml
openapi: 3.1.0
info:
  title: Yoon API
  version: 0.1.0
  summary: One API in front of African payment providers.
  description: |
    Yoon is a self-hosted payment gateway. Your application sends **intent** — collect
    5 000 XOF from this customer by Wave — and Yoon picks a configured provider, talks to
    it, verifies its callbacks and keeps a ledger.

    ## Conventions
    - **Authentication:** `Authorization: Bearer yk_…` on every `/v1` route except `/v1/about`.
    - **Money:** integer amounts in the currency's minor unit. XOF has none: 5 000 XOF is `5000`.
      Never floating point.
    - **Idempotency:** every `POST` requires an `Idempotency-Key` header. Retrying with the same
      key and body returns the original response (with `Idempotent-Replayed: true`). The same key
      with a different body is refused. Keys are kept for 24 hours.
    - **Errors:** [RFC 9457](https://www.rfc-editor.org/rfc/rfc9457) `application/problem+json`
      with a stable machine-readable `code`.
    - **Lists:** newest first, cursor-paginated with `starting_after` and `limit` (max 100).
    - **Phone numbers** are returned masked (`+22177***67`).
    - **Unknown outcomes:** if a provider does not answer, the payment stays `pending` on that
      provider and is never sent to another one — the customer may already be paying.
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
security:
  - apiKey: []
...
paths:
...
  /v1/payouts:
    post:
      tags: [Payouts]
      operationId: createPayout
      summary: Send a payout
      description: |
        Sent to exactly one provider, exactly once. If the provider's answer is lost the payout
        is `unknown` and Yoon resolves it through the provider's status API — it is never
        retried and never sent elsewhere, so it can never be paid twice.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePayoutRequest'
            example:
              amount: 10000
              currency: XOF
              country: SN
              method: wave
              recipient: { phone: '+221771234567' }
              reference: withdrawal_77
      responses:
        '201':
          description: Payout created.
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payout'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '422':
          description: No provider can send this payout (`no_provider_for_method`, `all_providers_unavailable`, `provider_not_configured`).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    get:
      tags: [Payouts]
      operationId: listPayouts
      summary: List payouts
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/PayoutStatus' } }
        - { name: reference, in: query, schema: { type: string } }
        - { name: needs_review, in: query, schema: { type: boolean }, description: Payouts whose outcome stayed unknown too long. }
        - $ref: '#/components/parameters/StartingAfter'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of payouts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayoutPage'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /admin/v1/dead-letters:
    get:
      tags: [Admin]
      operationId: adminListDeadLetters
      summary: Events whose delivery gave up
      security: [{ adminToken: [] }]
      parameters:
        - $ref: '#/components/parameters/StartingAfter'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Dead events, all applications.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AdminDisabled' }
  /admin/v1/dead-letters/{id}/replay:
    post:
      tags: [Admin]
      operationId: adminReplayDeadLetter
      summary: Deliver a dead event again
      security: [{ adminToken: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '202':
          description: Queued for delivery.
          content:
            application/json:
              schema:
                type: object
                required: [id, delivery_status]
                properties:
                  id: { type: string }
                  delivery_status: { type: string, const: pending }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AdminDisabled' }
        '422':
          description: '`not_dead` — the event is not dead.'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /admin/v1/payouts:
    get:
      tags: [Admin]
      operationId: adminListPayouts
      summary: Payouts needing review
      description: Payouts whose outcome stayed unknown past `YOON_SWEEP_PAYOUT_REVIEW_AFTER`, all applications.
      security: [{ adminToken: [] }]
      parameters:
        - { name: needs_review, in: query, schema: { type: boolean, default: true } }
        - $ref: '#/components/parameters/StartingAfter'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of payouts.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayoutPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AdminDisabled' }
  /admin/v1/payouts/{id}/resolve:
    post:
      tags: [Admin]
      operationId: adminResolvePayout
      summary: Settle a payout by hand
      description: |
        After checking with the provider directly. Goes through the payout state machine, the
        ledger and the application's events like any other change; the note is kept in the history.
      security: [{ adminToken: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, note]
              properties:
                status: { type: string, enum: [paid, failed] }
                note: { type: string, maxLength: 500, example: 'Confirmed on the PayDunya dashboard, tx 88123' }
      responses:
        '200':
          description: The resolved payout.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Payout' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AdminDisabled' }
        '422':
          description: '`already_final` or `illegal_transition`.'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: 'An application API key: `yk_…`. Create one with `yoon apps create <name>`.'
    adminToken:
      type: http
      scheme: bearer
      description: The operator token set in `YOON_ADMIN_TOKEN`.
  headers:
    IdempotentReplayed:
      description: Present and `true` when this response is a replay of an earlier request with the same key.
      schema: { type: string, enum: ['true'] }
  parameters:
...
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique per operation you intend, e.g. a UUID. Reuse it only to retry the same request.
      schema: { type: string, minLength: 1, maxLength: 255 }
...
    StartingAfter:
      name: starting_after
      in: query
      description: Cursor — the `next_cursor` of the previous page.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
    BadRequest:
      description: Invalid request (`invalid_request`).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid API key (`unauthorized`).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    IdempotencyConflict:
      description: |
        `idempotency_in_progress` — the first request with this key is still running; retry after
        `Retry-After` seconds. `idempotency_key_reused` — the key was used with a different body.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    AdminDisabled:
      description: The admin API is disabled (no `YOON_ADMIN_TOKEN`), or the resource does not exist.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Stable, machine-readable error code.
...
    PayoutPage:
      type: object
      required: [data, has_more, next_cursor]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Payout' } }
        has_more: { type: boolean }
        next_cursor: { $ref: '#/components/schemas/NextCursor' }
...
    EventPage:
      type: object
      required: [data, has_more, next_cursor]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Event' } }
        has_more: { type: boolean }
        next_cursor: { $ref: '#/components/schemas/NextCursor' }
    EventType:
      type: string
      enum: [payment.succeeded, payment.failed, payment.expired, refund.succeeded, refund.failed,
             payout.paid, payout.failed, payout.needs_review]
    DeliveryStatus:
      type: string
      enum: [pending, delivered, dead, no_endpoint]
      description: '`no_endpoint` — no webhook URL configured; the event is only available through the API.'
    EventPayload:
      type: object
      required: [id, object, type, created_at, data]
      description: The body POSTed to your webhook URL.
      properties:
        id: { type: string, example: evt_0199a3f0c2e47a1b9c3d5e6f7a8b9c0d, description: Deduplicate on this. }
        object: { type: string, const: event }
        type: { $ref: '#/components/schemas/EventType' }
        created_at: { type: string, format: date-time }
        data:
          type: object
          required: [object]
          properties:
            object:
              description: The payment, refund or payout as `GET` would return it at the time of the event.
              oneOf:
                - $ref: '#/components/schemas/Payment'
                - $ref: '#/components/schemas/Refund'
                - $ref: '#/components/schemas/Payout'
    Event:
      type: object
      required: [id, object, type, resource_type, resource_id, created_at, delivery, payload]
      properties:
        id: { type: string }
        object: { type: string, const: event }
        type: { $ref: '#/components/schemas/EventType' }
        resource_type: { type: string, enum: [payment, refund, payout] }
        resource_id: { type: string }
        created_at: { type: string, format: date-time }
        delivery:
          type: object
          required: [status, attempts]
          properties:
            status: { $ref: '#/components/schemas/DeliveryStatus' }
            attempts: { type: integer }
            next_attempt_at: { type: [string, 'null'], format: date-time }
            last_status_code: { type: [integer, 'null'] }
            last_error: { type: [string, 'null'] }
            delivered_at: { type: [string, 'null'], format: date-time }
        payload: { $ref: '#/components/schemas/EventPayload' }
    NextCursor:
      type: [string, 'null']
      description: Pass as `starting_after` to get the next page; null on the last page.
...
    PaymentStatus:
      type: string
      enum: [created, pending, succeeded, failed, expired]
      description: |
        `pending` — with the provider, waiting for the customer (or for the provider's answer).
        `succeeded` is final. `failed` and `expired` can still become `succeeded` if the provider
        later confirms the money arrived.
    RefundStatus:
      type: string
      enum: [created, pending, unknown, refunded, failed]
    PayoutStatus:
      type: string
      enum: [created, processing, unknown, paid, failed]
    Failure:
      type: [object, 'null']
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    Payment:
      type: object
      required: [id, object, status, amount, amount_refunded, currency, country, method, created_at, updated_at]
      properties:
        id: { type: string, example: pay_0199a3f0c2e47a1b9c3d5e6f7a8b9c0d }
        object: { type: string, const: payment }
        status: { $ref: '#/components/schemas/PaymentStatus' }
        amount: { type: integer, format: int64 }
        amount_refunded: { type: integer, format: int64 }
        currency: { type: string }
        country: { type: string }
        method:
          type: string
          example: wave
          description: >-
            The payment method. `any` for a hosted checkout whose customer has not chosen yet
            (kept a string so clients built against 0.1.0 keep working).
        checkout:
          type: string
          enum: [direct, hosted]
          description: '`hosted`: the customer picks the method on the Yoon page at `checkout_url`.'
        checkout_expires_at:
          type: [string, 'null']
          format: date-time
          description: Hosted checkout only. After this an unused checkout fails with `checkout_expired`.
        reference: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
        customer:
          type: [object, 'null']
          properties:
            phone: { type: [string, 'null'], example: '+22177***67' }
        provider: { type: [string, 'null'] }
        provider_reference: { type: [string, 'null'] }
        checkout_url:
          type: [string, 'null']
          description: |
            Redirect the customer here, if set. For a hosted checkout, always the Yoon page
            (`{YOON_PUBLIC_URL}/checkout/{id}?t=…`); it forwards the customer to the provider.
        instructions: { type: [string, 'null'], description: Push/USSD instruction to show the customer, if set. }
        routing_reason: { type: [string, 'null'], description: Why this provider — and any rejections before it. }
        failure: { $ref: '#/components/schemas/Failure' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Refund:
      type: object
      required: [id, object, payment_id, status, amount, currency, provider, created_at, updated_at]
      properties:
        id: { type: string, example: re_0199a3f0c2e47a1b9c3d5e6f7a8b9c0d }
        object: { type: string, const: refund }
        payment_id: { type: string }
        status: { $ref: '#/components/schemas/RefundStatus' }
        amount: { type: integer, format: int64 }
        currency: { type: string }
        reason: { type: [string, 'null'] }
        provider: { type: string }
        provider_reference: { type: [string, 'null'] }
        failure: { $ref: '#/components/schemas/Failure' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreatePayoutRequest:
      type: object
      required: [amount, currency, country, method, recipient]
      properties:
        amount: { type: integer, format: int64, minimum: 1 }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        country: { type: string, pattern: '^[A-Z]{2}$' }
        method: { type: string, maxLength: 32 }
        recipient:
          type: object
          description: 'A phone, or a PI-SPI payment alias (`pispi` provider): at least one.'
          properties:
            phone: { type: string }
            pi_alias: { type: string, maxLength: 64, description: "The recipient's PI-SPI payment alias. Required by the `pispi` provider." }
        reference: { type: string, maxLength: 255 }
        provider: { type: string, description: 'Optional: pin a configured provider.' }
    Payout:
      type: object
      required: [id, object, status, amount, currency, country, method, recipient, provider, needs_review, created_at, updated_at]
      properties:
        id: { type: string, example: po_0199a3f0c2e47a1b9c3d5e6f7a8b9c0d }
        object: { type: string, const: payout }
        status: { $ref: '#/components/schemas/PayoutStatus' }
        amount: { type: integer, format: int64 }
        currency: { type: string }
        country: { type: string }
        method: { type: string }
        reference: { type: [string, 'null'] }
        recipient:
          type: object
          required: [phone]
          properties:
            phone: { type: [string, 'null'], example: '+22177***67', description: Masked. Null when the payout went to a PI-SPI alias. }
        provider: { type: string }
        provider_reference: { type: [string, 'null'] }
        routing_reason: { type: [string, 'null'] }
        needs_review: { type: boolean, description: The outcome stayed unknown too long; a human should check with the provider. }
        failure: { $ref: '#/components/schemas/Failure' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...